Refactored tests into multiple files
